[Refactor] api 코드 컨벤션 통일, 파일 병합과 정리, teamId 인자로 받도록 변경, 필요 없어진 코드 제거
