+++ src/components/modal/NewDashboard.tsx
@@ -2,6 +2,8 @@ import { useState } from "react";
 import Input from "../input/Input";
 import Image from "next/image";
 import axios from "axios";
+import { TEAM_ID } from "@/constants/team";
+
 interface Dashboard {
   id: number;
   title: string;
@@ -35,7 +37,7 @@ export default function NewDashboard({
     try {
       setLoading(true);
       const response = await axios.post(
-        `${process.env.NEXT_PUBLIC_BASE_URL}/13-4/dashboards`,
+        `${process.env.NEXT_PUBLIC_BASE_URL}/${TEAM_ID}/dashboards`,
         payload,
         {
           headers: {
