+++ src/components/table/InviteRecords.tsx
@@ -21,7 +21,7 @@ const InviteRecords = ({ dashboardId }: { dashboardId: string }) => {
       try {
         const dashboardIdNumber = Number(dashboardId);
         const res = await axiosInstance.get(
-          apiRoutes.DashboardInvite(dashboardIdNumber),
+          apiRoutes.dashboardInvite(dashboardIdNumber),
           {
             params: {
               dashboardId,
@@ -55,7 +55,7 @@ const InviteRecords = ({ dashboardId }: { dashboardId: string }) => {
     if (!dashboardId) return;
     try {
       await axiosInstance.delete(
-        apiRoutes.DashboardInviteDelete(dashboardIdNumber, id)
+        apiRoutes.dashboardInviteDelete(dashboardIdNumber, id)
       );
       window.location.reload();
     } catch (error) {
