+++ src/pages/mypage.tsx
@@ -3,6 +3,7 @@ import HeaderMyPage from "@/components/gnb/HeaderDashboard";
 import SideMenu from "@/components/sideMenu/SideMenu";
 import ProfileCard from "@/components/card/Profile";
 import ChangePassword from "@/components/card/ChangePassword";
+import { TEAM_ID } from "@/constants/team";
 // import { getDashboards } from "@/api/dashboards";
 
 export default function MyPage() {
@@ -35,7 +36,7 @@ export default function MyPage() {
 
   return (
     <div className="flex">
-      <SideMenu teamId={"13-4"} dashboardList={[]} />
+      <SideMenu teamId={TEAM_ID} dashboardList={[]} />
       <div className="flex flex-col w-full">
         <HeaderMyPage variant="mypage" />
         <div className="flex flex-col items-center w-full mt-10">
